Dev v2.6

+++ src/main/java/com/pawar/inventory/app/repository/RolePermissionsRepository.java
@@ -8,12 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.pawar.inventory.app.model.RolePermissions;
 import com.pawar.inventory.app.model.RolePermissionsId;
 
-
-
 @Repository
 public interface RolePermissionsRepository extends CrudRepository<RolePermissions, RolePermissionsId> {
 
 	@Query("SELECT rp FROM RolePermissions rp WHERE rp.roleId = :roleId and rp.permissionId = :permissionId")
-	RolePermissions findRolesPermissionById(@Param("roleId") Integer roleId, @Param("permissionId") Long permissionId);
+	RolePermissions findRolesPermissionById(@Param("roleId") Integer roleId, @Param("permissionId") Integer permissionId);
 
 }
